+++ products/products/service.py
@@ -1,8 +1,8 @@
 from nameko.events import event_handler
 from nameko_grpc.entrypoint import Grpc
 
-from .proto.products_pb2_grpc import productsStub
-from .proto.products_pb2 import Product, Products
+from .products_pb2_grpc import productsStub
+from .products_pb2 import Product, Products
 
 from products import dependencies, schemas
